Try to workaround lack of gi split by also putting gi in host for cross.
Build errors when cross compiling seem related to finding non-host
libgirepository, so try adding it to host as well so the library can
come from there and the binaries can be used from build. This is where
we probably need to split the gobject-introspection package, but maybe
we can hack a workaround for now.