Add patch to so DLL name doesn't change when MSVC version changes
MSVC should maintain binary compatibility across versions anyway, as is
seen across the rest of conda-forge, so we don't want to break it
intentionally by changing the DLL name every time the version of MSVC
used changes. Luckily nothing on conda-forge outside of this
package seems to actually depend on the DLLs, since everything seems to
be using the Python bindings which only need the gobject-introspection
files also distributed as part of this package.